tavily-enterprise-demo/requirements.txt
# Tavily enterprise demo dependencies

# Core
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# CLI and logging
typer>=0.12.0
rich>=13.0.0

# Search and HTTP
tavily-python>=0.5.0
requests>=2.31.0

# LLM ranking
openai>=1.0.0
tenacity>=8.0.0

# Document content extraction (for verification)
pypdf>=4.0.0
openpyxl>=3.1.0

# Web UI
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0

# Testing
pytest>=8.0.0
